requirements.txt
fastapi
uvicorn
torch
pillow
timm
requests
python-multipart